Bug 366471 - Improve OSX scroll performance with drawText caching
Add caching to GC.drawText to prevent multiple calls with the same single
character from needed to perform expensive layout operations. Any call to
CG.drawText() that is rendering a single character on a transparent
background is now a cache candidate.

The cache is limited to 20 items and is specifically designed to work well
with classes such as WhitespaceCharacterPainter which frequently call
drawText() but operate with a relatively small set of characters.

Signed-off-by: Phillip Webb <pwebb@pivotal.com>