Bug 366471 - Improve OSX scroll performance by overriding isOpaque()
Update Composite.isOpaque() so that it delegates to protected method which
can be overridden by StyledText.

Prior to this commit, calls to StyledText.isOpaque() on OSX (with hidden scrollbars)
would always result in false. This meant that optimizations usually performed by
the OS could not happen.

There were two reasons that false was returned:

1) The `background` is usually set to `null`, meaning that the
   `background != null && background[3]` test always failed.

2) isObscured() always returns `true` due to the scrollbars that fade
   in and out as scrolling occurs. See Bug 292281 for details of why
   isObscured() is tested in isOpaque().

Signed-off-by: Phillip Webb <pwebb@pivotal.com>